ModuleType ceases to exist. Now the Module type is a String.
